response form logic fully migrated from forums to ResponseForm.js
