+++ src/pages/Forums.js
@@ -14,13 +14,13 @@ topic: '',
 body: ''
 });  
 
-const [responses, setResponses] = useState([]);
-const [responseFormInputs, updateResponseFormInputs] = useState({
-userName: 'defaultName',
-userImage: 'defaultUrl',
-category: 'defaultCat',
-body: ''
-})
+// const [responses, setResponses] = useState([]);
+// const [responseFormInputs, updateResponseFormInputs] = useState({
+// userName: 'defaultName',
+// userImage: 'defaultUrl',
+// category: 'defaultCat',
+// body: ''
+// })
 
 const getQuestions = async () => {
 try{
@@ -32,15 +32,15 @@ try{
 };
 };
 
-const getResponses = async () => {
-try{
-    const response = await fetch('http://localhost:3001/api/responses'); 
-    const rData = await response.json();
-    setResponses(rData)
-}catch(error){
-    console.error(error); 
-};
-};
+// const getResponses = async () => {
+// try{
+//     const response = await fetch('http://localhost:3001/api/responses'); 
+//     const rData = await response.json();
+//     setResponses(rData)
+// }catch(error){
+//     console.error(error); 
+// };
+// };
 
 useEffect(
 () => {
@@ -51,24 +51,24 @@ useEffect(
     )()
 }, [questions])
 
-useEffect(
-() => {
-    (
-    async function (){
-        await getResponses();
-    }
-    )()
-}, [responses])
+// useEffect(
+// () => {
+//     (
+//     async function (){
+//         await getResponses();
+//     }
+//     )()
+// }, [responses])
 
 const handleQuestionChange = (event) => {
 const updatedQuestionFormInputs = Object.assign({}, questionFormInputs, {[event.target.id]: event.target.value})
 updateQuestionFormInputs(updatedQuestionFormInputs);
 }
 
-const handleResponseChange = (event) => {
-const updatedResponseFormInputs = Object.assign({}, responseFormInputs, {[event.target.id]: event.target.value})
-updateResponseFormInputs(updatedResponseFormInputs);
-}
+// const handleResponseChange = (event) => {
+// const updatedResponseFormInputs = Object.assign({}, responseFormInputs, {[event.target.id]: event.target.value})
+// updateResponseFormInputs(updatedResponseFormInputs);
+// }
 
 const handleQuestionSubmit = async (event) => {
 event.preventDefault();
@@ -90,24 +90,24 @@ try{
 };
 };
 
-const handleResponseSubmit = async (event) => {
-event.preventDefault();
-console.log(responseFormInputs);
-try{
-    const response = await axios.post('http://localhost:3001/api/responses', responseFormInputs);
-    const createdResponse = response.data
-    console.log(`this is createdResponse ${createdResponse.body}`);
-    await updateResponseFormInputs({
-    userName: 'defaultName',
-    userImage: 'defaulturl',
-    category: 'defaultCat',
-    body: ''
-    });
-    await setResponses([createdResponse, ...responses])
-}catch(error){
-    console.error(error)
-};
-};
+// const handleResponseSubmit = async (event) => {
+// event.preventDefault();
+// console.log(responseFormInputs);
+// try{
+//     const response = await axios.post('http://localhost:3001/api/responses', responseFormInputs);
+//     const createdResponse = response.data
+//     console.log(`this is createdResponse ${createdResponse.body}`);
+//     await updateResponseFormInputs({
+//     userName: 'defaultName',
+//     userImage: 'defaulturl',
+//     category: 'defaultCat',
+//     body: ''
+//     });
+//     await setResponses([createdResponse, ...responses])
+// }catch(error){
+//     console.error(error)
+// };
+// };
 
 const handleQuestionDelete = async (event) => {
 event.preventDefault();
@@ -121,17 +121,17 @@ console.log(`look here ${event.target.id}`);
 };
 };
 
-const handleResponseDelete = async (event) => {
-event.preventDefault();
-event.persist();
-console.log(event);
-try {
-console.log(`look here ${event.target.id}`);
-    await axios.delete(`http://localhost:3001/api/responses/${event.target.id}`)
-} catch(error){
-    console.error(error)
-};
-};
+// const handleResponseDelete = async (event) => {
+// event.preventDefault();
+// event.persist();
+// console.log(event);
+// try {
+// console.log(`look here ${event.target.id}`);
+//     await axios.delete(`http://localhost:3001/api/responses/${event.target.id}`)
+// } catch(error){
+//     console.error(error)
+// };
+// };
 
 const [questionUpdate, setQuestionUpdate] = useState({}) 
 
@@ -244,7 +244,7 @@ return (
 
 
 
-    <h1>This is the Forums component - Response form</h1>
+    {/* <h1>This is the Forums component - Response form</h1>
     <form onSubmit={handleResponseSubmit}>
         <h3>Post Your Response</h3>
         <label htmlFor='body'>Response</label>
@@ -255,9 +255,9 @@ return (
         value={responseFormInputs.body}
         /><br />
         <input type='submit' value='Post Response' />
-    </form>
+    </form> */}
 
-    <div>
+    {/* <div>
         {
         responses.map(res =>{
             return(
@@ -270,7 +270,7 @@ return (
             )
         })
         }
-    </div>
+    </div> */}
